add dynamic array code

+++ src/sema/ast.rs
@@ -446,7 +446,7 @@ pub enum Expression {
         loc: program::Loc,
         ty: Type,
         length: Box<Expression>,
-        init: Option<Vec<u8>>,
+        init: Option<Vec<u64>>,
     },
 
     StorageArrayLength {
